% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_gwalpha.R
\name{sum_gwalpha}
\alias{sum_gwalpha}
\title{Counting the Number of GWalpha below Criterion}
\usage{
sum_gwalpha(bw, crit, ...)
}
\arguments{
\item{bw}{bandwidth for weighting function}

\item{crit}{predetermined criterion for reliability level}

\item{...}{additional arguments passed on to \code{\link[=gwalpha]{gwalpha()}}}
}
\value{
the number of responses where the upper confidence bound of GWalpha is below
\code{crit}.
}
\description{
an internal function for counting how many responses have an upper confidence
bound below a user-specified criterion for reliability level.
}
